JPA repository,Dependency injection and database seeding

+++ src/main/java/com/webproject/course/entities/User.java
@@ -11,9 +11,10 @@ import java.util.Objects;
 public class User implements Serializable {
     private static final long serialVersionUID = 1L;
 
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
     private String name;
     private String email;
     private String phone;
@@ -21,7 +22,7 @@ public class User implements Serializable {
 
     public User(){}
 
-    public User(long id, String name, String email, String phone, String password) {
+    public User(Integer id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -33,7 +34,7 @@ public class User implements Serializable {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
